requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Deep learning and RL
torch>=2.0.0
torchvision>=0.15.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0

# Interpretability libraries
shap>=0.42.0
lime>=0.2.0
captum>=0.6.0

# Visualization and demos
streamlit>=1.25.0
gradio>=3.40.0

# Testing and quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Optional: tracking and serving
wandb>=0.15.0
mlflow>=2.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
